+++ src/App.tsx
@@ -1,7 +1,6 @@
 // import React, { useRef } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Login from './pages/Login/Login';
-import LoginTest from './pages/Login/InstagramLogin';
 import './styles/reset.scss';
 
 function App() {
@@ -14,7 +13,6 @@ function App() {
       <Routes>
         {/* <Route path='/' element= */}
         <Route path='/login' element={<Login />} />
-        <Route path='/logintest' element={<LoginTest />} />
       </Routes>
     </BrowserRouter>
   );
