+++ src/styles/_variabled.scss
@@ -1,4 +1,6 @@
 $ACCENT_COLOR: #9564C5;
+$ACCENT_LIGHT_COLOR: #AB7ED2;
+$ACCENT_DARK_COLOR: #7E43B1;
 $LIGHT_THEME: (
     bg-color: #FFF,
     sub-bg-color: #F3F3F3,
